Switch to AGPL license
After giving this some thought for a few weeks I came to the following conclusion.

Basically, since Armored Bits is a very limited scope network game that I want to open source to a) be a public example of my Haskell programming abilities, and b) be examined and possibly contributed to by the community that grows around the game, I wanted to protect it behind a license that would stop an entity from outright redistributing my hard work and profiting from it without contributing back.

Armored Bits is not a generally useful piece of software, so given that context I believe that as an independent game developer this is the right choice for an open source network game project.